Add View OrderDetails and Print Receipt

+++ app/src/main/java/io/ramani/ramaniStationary/data/history/models/response/Order.kt
@@ -15,5 +15,5 @@ data class Order(
     val paymentStatus: String = "",
     val payments: List<Any> = listOf(),
     val qtyRewardItems: List<Any> = listOf(),
-    val totalCost: Int = 0
+    val totalCost: Double = 0.0
 )
